+++ refena/lib/src/provider/types/future_provider.dart
@@ -3,6 +3,7 @@ import 'package:refena/src/async_value.dart';
 import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
+import 'package:refena/src/provider/provider_accessor.dart';
 import 'package:refena/src/provider/types/future_family_provider.dart';
 import 'package:refena/src/proxy_ref.dart';
 import 'package:refena/src/ref.dart';
@@ -18,10 +19,11 @@ import 'package:refena/src/ref.dart';
 /// Example use cases:
 /// - fetch static data from an API (that does not change)
 /// - fetch device information (that does not change)
-class FutureProvider<T>
-    extends BaseWatchableProvider<FutureProviderNotifier<T>, AsyncValue<T>>
+class FutureProvider<T> extends BaseWatchableProvider<FutureProvider<T>,
+        FutureProviderNotifier<T>, AsyncValue<T>>
     with
-        ProviderSelectMixin<FutureProviderNotifier<T>, AsyncValue<T>>
+        ProviderSelectMixin<FutureProvider<T>, FutureProviderNotifier<T>,
+            AsyncValue<T>>
     implements
         RebuildableProvider<FutureProviderNotifier<T>, AsyncValue<T>,
             Future<T>> {